test: point controller, service 분리하도록 테스트 케이스 전면 수정
- controller, service를 분리하면 관심사가 나누어져 더 쉽게 비즈니스 로직만 볼 수 있음.
